mangrove height
