+++ src/components/widget-custom-label/index.js
@@ -1,0 +1,1 @@
+export { default } from './component';
